inventory_browser: fix bug where itemsIds was overriden by an empty array
which seem to be due to the duplicated calls to setupPagination,
while the fetching flag was handled by PaginatedItems
